refactored abstract methods
